AK+FileManager: Move out human_readable_size to AK::NumberFormat
